Move pkg/httpfs to internal/httpfs
