+++ cmd/query/app/ui/assets.go
@@ -10,7 +10,7 @@ import (
 	"go.uber.org/zap"
 
 	"github.com/jaegertracing/jaeger/internal/gzipfs"
-	"github.com/jaegertracing/jaeger/pkg/httpfs"
+	"github.com/jaegertracing/jaeger/internal/httpfs"
 )
 
 //go:embed actual/*
